refactor(tao)!: accept the window API v2 only through the AWT-free clone
Compose's own `androidx.compose.ui.window.v2` types are no longer accepted by
`DecoratedWindow` / `DecoratedDialog` / `HostedWindow` / `HostedDialog` or
the `NucleusWindowHost` / `NucleusDialogHost` surfaces. On Tao that surface
could only ever be half-working — every scoped geometry provider (including
the ones `requestSize` / `requestPosition` build internally) and `requestScreen`
were accepted, logged and dropped, because Compose's scope needs a displayable
`java.awt.Window`. An API that silently ignores part of its contract is worse
than one that does not exist; the supported v2 surface is the clone,
`dev.nucleusframework.window.tao.v2`, where everything is applied and
migrating is one import.

Removed: `ComposeWindowV2Bridge`, the `ComposeWindowV2Access` friend-package
accessor, the Compose-typed `DecoratedWindow` / `DecoratedDialog` overloads,
`inspectableWindowBounds` / `requestInspectableBounds`,
`rememberSyncedWindowState` / `rememberSyncedDialogState`, the matching
nucleus-application overloads, adapters and host methods, and their tests.
The geometry helpers the clone shared with that bridge move into
`NucleusWindowV2Bridge`. `examples/tao-demo` and the host tests use the clone.
