+++ CLAUDE.md
@@ -68,8 +68,8 @@ Published releases are `2.5.x` (latest tag `v2.5.0`). Do not treat `IDEAL_API.md
 - Native modules use platform-specific JNI implementations — test on each OS
 - Plugin is published via included build in `plugin-build/`
 - Version catalog is the source of truth for all dependency versions
-- **Compose window API v2**: Compose 1.12's `androidx.compose.ui.window.v2` types are hard-wired to AWT (`Screen` wraps a `GraphicsDevice`, `WindowGeometryProviderScope` takes a displayable `java.awt.Window`), so on Tao every scoped geometry provider and `requestScreen` are inert — accepted, logged, dropped. `dev.nucleusframework.window.tao.v2` is a member-for-member AWT-free clone of that package backed by `TaoMonitors` + `TaoWindow`: migrating is a single import change, and deleting the package restores the upstream import if JetBrains decouples its own types. Multi-monitor geometry comes from `TaoMonitors` (`EnumDisplayMonitors` / `NSScreen.screens` / GDK), never `GraphicsEnvironment`
-- **No reflection**: runtime modules must stay GraalVM native-image compatible, so reflection is not an acceptable implementation tool — not even with a graceful fallback. Reach for a static bridge instead (e.g. a friend-package Java accessor like `androidx.compose.ui.window.v2.ComposeWindowV2Access`, which reads Kotlin `internal` members through their `$ui`-mangled JVM names), a public API of our own, or a plugin bytecode transform. A feature that can only be built reflectively is a feature we do not ship: document the gap and offer a working alternative
+- **Compose window API v2**: Compose 1.12's `androidx.compose.ui.window.v2` types are hard-wired to AWT (`Screen` wraps a `GraphicsDevice`, `WindowGeometryProviderScope` takes a displayable `java.awt.Window`), so they are **not accepted** by any Nucleus window API — a half-working surface (scoped providers and `requestScreen` inert) is worse than none. The supported v2 surface is `dev.nucleusframework.window.tao.v2`, a member-for-member AWT-free clone backed by `TaoMonitors` + `TaoWindow`: migrating from the Compose package is a single import change, and deleting the clone restores the upstream import if JetBrains decouples its own types. Multi-monitor geometry comes from `TaoMonitors` (`EnumDisplayMonitors` / `NSScreen.screens` / GDK), never `GraphicsEnvironment`
+- **No reflection**: runtime modules must stay GraalVM native-image compatible, so reflection is not an acceptable implementation tool — not even with a graceful fallback. Reach for a static bridge instead (e.g. a friend-package Java accessor like `androidx.compose.ui.draganddrop.TaoTransferableAccess`, which reads Kotlin `internal` members through their `$ui`-mangled JVM names), a public API of our own, or a plugin bytecode transform. A feature that can only be built reflectively is a feature we do not ship: document the gap and offer a working alternative
 - **Public API freeze**: root `build.gradle.kts` applies kotlinx binary-compatibility-validator + `explicitApi()` to every non-example module. Baselines live in `<module>/api/<module>.api`. After intentional public API changes run `./gradlew apiDump` and commit the dump; `apiCheck` (wired into `check` / `preMerge`) fails on accidental ABI drift. Exception: `decorated-window-jewel` (JVM 25) is ignored by BCV until ASM supports class-file 69 — still uses `explicitApi()`. Helper: `scripts/fix-explicit-api.py` for mechanical visibility/return-type fixes from kotlinc diagnostics.
 - **KDoc on public API**: `UndocumentedPublicClass` / `UndocumentedPublicFunction` are enforced by detekt (`detekt` is wired into `check` / `preMerge`). Pre-existing gaps are grandfathered in per-module `<module>/detekt-baseline.xml` files — any *new* undocumented public class or function fails the build. Do not regenerate a baseline to silence a new finding; write the KDoc. `UndocumentedPublicProperty` stays off because the generated icon/symbol catalogs (`sf-symbols`, `freedesktop-icons`) would swamp it
 - **Logging**: `java.util.logging` is the single facade for every runtime module — no SLF4J dependency forced on consumers, no raw `println` / `System.err` in `src/main`. Logger names must be the fully-qualified class name (or an explicit `dev.nucleusframework.*` string) so the whole framework sits under one JUL namespace. `allowNucleusRuntimeLogging = true` is an opt-in convenience that raises the `dev.nucleusframework` logger to `nucleusLoggingLevel` and attaches a colored console handler; apps that configure JUL themselves (`logging.properties`, `jul-to-slf4j`) leave it `false` and Nucleus never touches the JUL configuration
