+++ decorated-window-tao/api/decorated-window-tao.api
@@ -178,11 +178,6 @@ public final class dev/nucleusframework/window/tao/ComposableSingletons$NativeVi
 	public final fun getLambda$1447510722$Nucleus_decorated_window_tao ()Lkotlin/jvm/functions/Function2;
 }
 
-public final class dev/nucleusframework/window/tao/ComposeWindowV2BridgeKt {
-	public static final fun rememberSyncedDialogState (Landroidx/compose/ui/window/v2/DialogState;ZLandroidx/compose/runtime/Composer;I)Landroidx/compose/ui/window/DialogState;
-	public static final fun rememberSyncedWindowState (Landroidx/compose/ui/window/v2/WindowState;ZLandroidx/compose/runtime/Composer;I)Landroidx/compose/ui/window/WindowState;
-}
-
 public final class dev/nucleusframework/window/tao/D3D11TestTextureProducer : java/lang/AutoCloseable {
 	public static final field $stable I
 	public static final field Companion Ldev/nucleusframework/window/tao/D3D11TestTextureProducer$Companion;
@@ -202,10 +197,6 @@ public final class dev/nucleusframework/window/tao/DecoratedDialogKt {
 	public static final fun DecoratedDialog (Ldev/nucleusframework/window/tao/ApplicationScope;Lkotlin/jvm/functions/Function0;Landroidx/compose/ui/window/DialogState;ZLjava/lang/String;Landroidx/compose/ui/graphics/painter/Painter;ZZZLkotlin/jvm/functions/Function1;Lkotlin/jvm/functions/Function1;Landroidx/compose/runtime/CompositionLocalContext;Lkotlin/jvm/functions/Function3;Landroidx/compose/runtime/Composer;III)V
 }
 
-public final class dev/nucleusframework/window/tao/DecoratedDialogV2Kt {
-	public static final fun DecoratedDialog-imfDCbw (Ldev/nucleusframework/window/tao/ApplicationScope;Lkotlin/jvm/functions/Function0;Landroidx/compose/ui/window/v2/DialogState;ZLjava/lang/String;Landroidx/compose/ui/graphics/painter/Painter;ZZZJJLkotlin/jvm/functions/Function1;Lkotlin/jvm/functions/Function1;Landroidx/compose/runtime/CompositionLocalContext;Lkotlin/jvm/functions/Function3;Landroidx/compose/runtime/Composer;III)V
-}
-
 public final class dev/nucleusframework/window/tao/DecoratedWindowComposableKt {
 	public static final fun DecoratedWindow-sYvZbhs (Ldev/nucleusframework/window/tao/ApplicationScope;Lkotlin/jvm/functions/Function0;Landroidx/compose/ui/window/WindowState;Ljava/lang/String;Landroidx/compose/ui/graphics/painter/Painter;Landroidx/compose/ui/unit/DpSize;ZZZZZZZZLdev/nucleusframework/window/tao/TaoWindow;Lkotlin/jvm/functions/Function1;Lkotlin/jvm/functions/Function1;ZLdev/nucleusframework/window/tao/MacOSStyle;ZLandroidx/compose/runtime/CompositionLocalContext;ZZZZLkotlin/jvm/functions/Function3;Landroidx/compose/runtime/Composer;IIII)V
 }
@@ -217,10 +208,6 @@ public final class dev/nucleusframework/window/tao/DecoratedWindowKt {
 public final class dev/nucleusframework/window/tao/DecoratedWindowNucleusV2Kt {
 	public static final fun DecoratedDialog-imfDCbw (Ldev/nucleusframework/window/tao/ApplicationScope;Lkotlin/jvm/functions/Function0;Ldev/nucleusframework/window/tao/v2/DialogState;ZLjava/lang/String;Landroidx/compose/ui/graphics/painter/Painter;ZZZJJLkotlin/jvm/functions/Function1;Lkotlin/jvm/functions/Function1;Landroidx/compose/runtime/CompositionLocalContext;Lkotlin/jvm/functions/Function3;Landroidx/compose/runtime/Composer;III)V
 	public static final fun DecoratedWindow-INFUufI (Ldev/nucleusframework/window/tao/ApplicationScope;Lkotlin/jvm/functions/Function0;Ldev/nucleusframework/window/tao/v2/WindowState;Ljava/lang/String;Landroidx/compose/ui/graphics/painter/Painter;JJZZZZZZZZLdev/nucleusframework/window/tao/TaoWindow;Lkotlin/jvm/functions/Function1;Lkotlin/jvm/functions/Function1;ZLdev/nucleusframework/window/tao/MacOSStyle;ZLandroidx/compose/runtime/CompositionLocalContext;ZZZZLkotlin/jvm/functions/Function3;Landroidx/compose/runtime/Composer;IIII)V
-}
-
-public final class dev/nucleusframework/window/tao/DecoratedWindowV2Kt {
-	public static final fun DecoratedWindow-INFUufI (Ldev/nucleusframework/window/tao/ApplicationScope;Lkotlin/jvm/functions/Function0;Landroidx/compose/ui/window/v2/WindowState;Ljava/lang/String;Landroidx/compose/ui/graphics/painter/Painter;JJZZZZZZZZLdev/nucleusframework/window/tao/TaoWindow;Lkotlin/jvm/functions/Function1;Lkotlin/jvm/functions/Function1;ZLdev/nucleusframework/window/tao/MacOSStyle;ZLandroidx/compose/runtime/CompositionLocalContext;ZZZZLkotlin/jvm/functions/Function3;Landroidx/compose/runtime/Composer;IIII)V
 }
 
 public final class dev/nucleusframework/window/tao/DefaultWindowExceptionHandlerFactory : dev/nucleusframework/window/tao/WindowExceptionHandlerFactory {
@@ -245,15 +232,6 @@ public final class dev/nucleusframework/window/tao/DmaBufTestTextureProducer$Com
 	public static synthetic fun create$default (Ldev/nucleusframework/window/tao/DmaBufTestTextureProducer$Companion;IIIILjava/lang/Object;)Ldev/nucleusframework/window/tao/DmaBufTestTextureProducer;
 	public final fun createYuv (IILdev/nucleusframework/window/tao/NucleusYuvFormat;Ldev/nucleusframework/window/tao/NucleusYuvColorSpace;)Ldev/nucleusframework/window/tao/DmaBufTestTextureProducer;
 	public static synthetic fun createYuv$default (Ldev/nucleusframework/window/tao/DmaBufTestTextureProducer$Companion;IILdev/nucleusframework/window/tao/NucleusYuvFormat;Ldev/nucleusframework/window/tao/NucleusYuvColorSpace;ILjava/lang/Object;)Ldev/nucleusframework/window/tao/DmaBufTestTextureProducer;
-}
-
-public final class dev/nucleusframework/window/tao/InspectableWindowBoundsKt {
-	public static final fun inspectableWindowBounds-8P0U83o (Landroidx/compose/ui/unit/DpSize;Landroidx/compose/ui/window/WindowPosition;)Landroidx/compose/ui/window/v2/WindowBoundsProvider;
-	public static synthetic fun inspectableWindowBounds-8P0U83o$default (Landroidx/compose/ui/unit/DpSize;Landroidx/compose/ui/window/WindowPosition;ILjava/lang/Object;)Landroidx/compose/ui/window/v2/WindowBoundsProvider;
-	public static final fun requestInspectableBounds-veQNT8c (Landroidx/compose/ui/window/v2/DialogState;Landroidx/compose/ui/unit/DpSize;Landroidx/compose/ui/window/WindowPosition;)V
-	public static final fun requestInspectableBounds-veQNT8c (Landroidx/compose/ui/window/v2/WindowState;Landroidx/compose/ui/unit/DpSize;Landroidx/compose/ui/window/WindowPosition;)V
-	public static synthetic fun requestInspectableBounds-veQNT8c$default (Landroidx/compose/ui/window/v2/DialogState;Landroidx/compose/ui/unit/DpSize;Landroidx/compose/ui/window/WindowPosition;ILjava/lang/Object;)V
-	public static synthetic fun requestInspectableBounds-veQNT8c$default (Landroidx/compose/ui/window/v2/WindowState;Landroidx/compose/ui/unit/DpSize;Landroidx/compose/ui/window/WindowPosition;ILjava/lang/Object;)V
 }
 
 public final class dev/nucleusframework/window/tao/MacOSStyle : java/lang/Enum {
