+++ decorated-window-tao/src/test/kotlin/dev/nucleusframework/window/tao/TaoSceneTestBatteryDriftTest.kt
@@ -108,8 +108,6 @@ class TaoSceneTestBatteryDriftTest {
                 "pure-Kotlin portal parent / xdg_foreign handle formatting, no scene",
             dev.nucleusframework.window.ChromeLogicTest::class.java to
                 "unit tests for chrome helpers; no ComposeScene",
-            ComposeWindowV2BridgeTest::class.java to
-                "pure Compose window API v1↔v2 state mapping, no ComposeScene",
             NucleusWindowV2BridgeTest::class.java to
                 "pure state mapping + geometry provider evaluation, no ComposeScene",
             TaoMonitorsTest::class.java to
