+++ nucleus-application/src/main/kotlin/dev/nucleusframework/application/DecoratedDialog.kt
@@ -10,7 +10,6 @@ import androidx.compose.ui.unit.DpSize
 import androidx.compose.ui.window.DialogState
 import androidx.compose.ui.window.rememberDialogState
 import dev.nucleusframework.application.internal.TaoDecoratedDialogAdapter
-import androidx.compose.ui.window.v2.DialogState as DialogStateV2
 import dev.nucleusframework.window.tao.v2.DialogState as NucleusDialogState
 
 /**
@@ -81,94 +80,6 @@ public fun DecoratedDialog(
         resizable = resizable,
         enabled = enabled,
         focusable = focusable,
-        onPreviewKeyEvent = onPreviewKeyEvent,
-        onKeyEvent = onKeyEvent,
-        content = content,
-    )
-}
-
-/**
- * [DecoratedDialog] overload for Compose Multiplatform 1.12's experimental
- * dialog API v2.
- *
- * [state] has no default so `DecoratedDialog(onCloseRequest) { }` still
- * resolves to the v1 overload.
- *
- * `requestScreen` / `screenId` are not applied on Tao (primary work area
- * only).
- */
-@ExperimentalComposeUiApi
-@Suppress("FunctionNaming", "LongParameterList")
-@Composable
-public fun NucleusApplicationScope.DecoratedDialog(
-    onCloseRequest: () -> Unit,
-    state: DialogStateV2,
-    visible: Boolean = true,
-    title: String = "",
-    icon: Painter? = null,
-    resizable: Boolean = false,
-    enabled: Boolean = true,
-    focusable: Boolean = true,
-    minSize: DpSize = DpSize.Unspecified,
-    maxSize: DpSize = DpSize.Unspecified,
-    onPreviewKeyEvent: (KeyEvent) -> Boolean = { false },
-    onKeyEvent: (KeyEvent) -> Boolean = { false },
-    content: @Composable NucleusDecoratedDialogScope.() -> Unit,
-) {
-    when (this) {
-        is TaoNucleusApplicationScope ->
-            TaoDecoratedDialogAdapter.DialogV2(
-                scope = this,
-                onCloseRequest = onCloseRequest,
-                state = state,
-                visible = visible,
-                title = title,
-                icon = icon,
-                resizable = resizable,
-                enabled = enabled,
-                focusable = focusable,
-                minSize = minSize,
-                maxSize = maxSize,
-                onPreviewKeyEvent = onPreviewKeyEvent,
-                onKeyEvent = onKeyEvent,
-                content = content,
-            )
-    }
-}
-
-/**
- * Receiver-less [DecoratedDialog] for Compose window API v2. See the
- * [NucleusApplicationScope] overload.
- */
-@ExperimentalComposeUiApi
-@Suppress("FunctionNaming", "LongParameterList")
-@Composable
-public fun DecoratedDialog(
-    onCloseRequest: () -> Unit,
-    state: DialogStateV2,
-    visible: Boolean = true,
-    title: String = "",
-    icon: Painter? = null,
-    resizable: Boolean = false,
-    enabled: Boolean = true,
-    focusable: Boolean = true,
-    minSize: DpSize = DpSize.Unspecified,
-    maxSize: DpSize = DpSize.Unspecified,
-    onPreviewKeyEvent: (KeyEvent) -> Boolean = { false },
-    onKeyEvent: (KeyEvent) -> Boolean = { false },
-    content: @Composable NucleusDecoratedDialogScope.() -> Unit,
-) {
-    LocalNucleusApplicationScope.current.DecoratedDialog(
-        onCloseRequest = onCloseRequest,
-        state = state,
-        visible = visible,
-        title = title,
-        icon = icon,
-        resizable = resizable,
-        enabled = enabled,
-        focusable = focusable,
-        minSize = minSize,
-        maxSize = maxSize,
         onPreviewKeyEvent = onPreviewKeyEvent,
         onKeyEvent = onKeyEvent,
         content = content,
