+++ nucleus-application/src/main/kotlin/dev/nucleusframework/application/DecoratedWindow.kt
@@ -10,7 +10,6 @@ import androidx.compose.ui.unit.DpSize
 import androidx.compose.ui.window.WindowState
 import androidx.compose.ui.window.rememberWindowState
 import dev.nucleusframework.application.internal.TaoDecoratedWindowAdapter
-import androidx.compose.ui.window.v2.WindowState as WindowStateV2
 import dev.nucleusframework.window.tao.v2.WindowState as NucleusWindowState
 
 /**
@@ -174,138 +173,6 @@ public fun DecoratedWindow(
         nativeContextMenu = nativeContextMenu,
         hiddenFromDock = hiddenFromDock,
         minimumSize = minimumSize,
-        onPreviewKeyEvent = onPreviewKeyEvent,
-        onKeyEvent = onKeyEvent,
-        transparent = transparent,
-        clickThrough = clickThrough,
-        visibleOnAllWorkspaces = visibleOnAllWorkspaces,
-        forceX11 = forceX11,
-        alwaysOnBottom = alwaysOnBottom,
-        content = content,
-    )
-}
-
-/**
- * [DecoratedWindow] overload for Compose Multiplatform 1.12's experimental
- * window API v2.
- *
- * [state] has no default so `DecoratedWindow(onCloseRequest) { }` still
- * resolves to the v1 overload.
- *
- * `requestScreen` / `screenId` are not applied on Tao (primary work area
- * only).
- */
-@ExperimentalComposeUiApi
-@Suppress("FunctionNaming", "LongParameterList")
-@Composable
-public fun NucleusApplicationScope.DecoratedWindow(
-    onCloseRequest: () -> Unit,
-    state: WindowStateV2,
-    visible: Boolean = true,
-    title: String = "",
-    icon: Painter? = null,
-    resizable: Boolean = true,
-    enabled: Boolean = true,
-    focusable: Boolean = true,
-    alwaysOnTop: Boolean = false,
-    undecorated: Boolean = false,
-    popupFor: NucleusWindow? = null,
-    nativePopupLayers: Boolean = false,
-    nativeContextMenu: Boolean = false,
-    hiddenFromDock: Boolean = false,
-    minSize: DpSize = DpSize.Unspecified,
-    maxSize: DpSize = DpSize.Unspecified,
-    onPreviewKeyEvent: (KeyEvent) -> Boolean = { false },
-    onKeyEvent: (KeyEvent) -> Boolean = { false },
-    transparent: Boolean = false,
-    clickThrough: Boolean = false,
-    visibleOnAllWorkspaces: Boolean = false,
-    forceX11: Boolean = false,
-    alwaysOnBottom: Boolean = false,
-    content: @Composable NucleusDecoratedWindowScope.() -> Unit,
-) {
-    when (this) {
-        is TaoNucleusApplicationScope ->
-            TaoDecoratedWindowAdapter.WindowV2(
-                scope = this,
-                onCloseRequest = onCloseRequest,
-                state = state,
-                visible = visible,
-                title = title,
-                icon = icon,
-                resizable = resizable,
-                enabled = enabled,
-                focusable = focusable,
-                alwaysOnTop = alwaysOnTop,
-                undecorated = undecorated,
-                transparent = transparent,
-                clickThrough = clickThrough,
-                visibleOnAllWorkspaces = visibleOnAllWorkspaces,
-                forceX11 = forceX11,
-                alwaysOnBottom = alwaysOnBottom,
-                popupFor = popupFor,
-                nativePopupLayers = nativePopupLayers,
-                nativeContextMenu = nativeContextMenu,
-                hiddenFromDock = hiddenFromDock,
-                minSize = minSize,
-                maxSize = maxSize,
-                onPreviewKeyEvent = onPreviewKeyEvent,
-                onKeyEvent = onKeyEvent,
-                content = content,
-            )
-    }
-}
-
-/**
- * Receiver-less [DecoratedWindow] for Compose window API v2. See the
- * [NucleusApplicationScope] overload.
- */
-@ExperimentalComposeUiApi
-@Suppress("FunctionNaming", "LongParameterList")
-@Composable
-public fun DecoratedWindow(
-    onCloseRequest: () -> Unit,
-    state: WindowStateV2,
-    visible: Boolean = true,
-    title: String = "",
-    icon: Painter? = null,
-    resizable: Boolean = true,
-    enabled: Boolean = true,
-    focusable: Boolean = true,
-    alwaysOnTop: Boolean = false,
-    undecorated: Boolean = false,
-    popupFor: NucleusWindow? = null,
-    nativePopupLayers: Boolean = false,
-    nativeContextMenu: Boolean = false,
-    hiddenFromDock: Boolean = false,
-    minSize: DpSize = DpSize.Unspecified,
-    maxSize: DpSize = DpSize.Unspecified,
-    onPreviewKeyEvent: (KeyEvent) -> Boolean = { false },
-    onKeyEvent: (KeyEvent) -> Boolean = { false },
-    transparent: Boolean = false,
-    clickThrough: Boolean = false,
-    visibleOnAllWorkspaces: Boolean = false,
-    forceX11: Boolean = false,
-    alwaysOnBottom: Boolean = false,
-    content: @Composable NucleusDecoratedWindowScope.() -> Unit,
-) {
-    LocalNucleusApplicationScope.current.DecoratedWindow(
-        onCloseRequest = onCloseRequest,
-        state = state,
-        visible = visible,
-        title = title,
-        icon = icon,
-        resizable = resizable,
-        enabled = enabled,
-        focusable = focusable,
-        alwaysOnTop = alwaysOnTop,
-        undecorated = undecorated,
-        popupFor = popupFor,
-        nativePopupLayers = nativePopupLayers,
-        nativeContextMenu = nativeContextMenu,
-        hiddenFromDock = hiddenFromDock,
-        minSize = minSize,
-        maxSize = maxSize,
         onPreviewKeyEvent = onPreviewKeyEvent,
         onKeyEvent = onKeyEvent,
         transparent = transparent,
