+++ nucleus-application/src/main/kotlin/dev/nucleusframework/application/NucleusWindowHost.kt
@@ -14,12 +14,8 @@ import androidx.compose.ui.window.DialogState
 import androidx.compose.ui.window.WindowState
 import androidx.compose.ui.window.rememberDialogState
 import androidx.compose.ui.window.rememberWindowState
-import dev.nucleusframework.window.tao.rememberSyncedDialogState
 import dev.nucleusframework.window.tao.rememberSyncedNucleusDialogState
 import dev.nucleusframework.window.tao.rememberSyncedNucleusWindowState
-import dev.nucleusframework.window.tao.rememberSyncedWindowState
-import androidx.compose.ui.window.v2.DialogState as DialogStateV2
-import androidx.compose.ui.window.v2.WindowState as WindowStateV2
 import dev.nucleusframework.window.tao.v2.DialogState as NucleusDialogState
 import dev.nucleusframework.window.tao.v2.WindowState as NucleusWindowState
 
@@ -83,66 +79,6 @@ public fun interface NucleusWindowHost {
         alwaysOnBottom: Boolean,
         content: @Composable NucleusDecoratedWindowScope.() -> Unit,
     )
-
-    /**
-     * Opens a window driven by Compose Multiplatform 1.12's experimental
-     * window API v2.
-     *
-     * Default implementation converts [state] to v1 and calls
-     * [Window] so existing themed hosts keep their chrome. `maxSize` is
-     * v2-only and is dropped on that fallback. Override to keep max-size
-     * or custom v2 chrome. `requestScreen` / `screenId` are not applied
-     * on Tao (primary work area only).
-     */
-    @ExperimentalComposeUiApi
-    @Suppress("UnusedParameter")
-    @Composable
-    public fun Window(
-        onCloseRequest: () -> Unit,
-        state: WindowStateV2,
-        visible: Boolean,
-        title: String,
-        icon: Painter?,
-        resizable: Boolean,
-        enabled: Boolean,
-        focusable: Boolean,
-        alwaysOnTop: Boolean,
-        undecorated: Boolean,
-        popupFor: NucleusWindow?,
-        nativePopupLayers: Boolean,
-        nativeContextMenu: Boolean,
-        hiddenFromDock: Boolean,
-        minSize: DpSize,
-        maxSize: DpSize,
-        onPreviewKeyEvent: (KeyEvent) -> Boolean,
-        onKeyEvent: (KeyEvent) -> Boolean,
-        alwaysOnBottom: Boolean,
-        content: @Composable NucleusDecoratedWindowScope.() -> Unit,
-    ) {
-        val v1 = rememberSyncedWindowState(state, visible)
-        Window(
-            onCloseRequest = onCloseRequest,
-            state = v1,
-            visible = visible,
-            title = title,
-            icon = icon,
-            resizable = resizable,
-            enabled = enabled,
-            focusable = focusable,
-            alwaysOnTop = alwaysOnTop,
-            undecorated = undecorated,
-            popupFor = popupFor,
-            nativePopupLayers = nativePopupLayers,
-            nativeContextMenu = nativeContextMenu,
-            hiddenFromDock = hiddenFromDock,
-            minimumSize =
-                if (minSize.width.isSpecified && minSize.height.isSpecified) minSize else null,
-            onPreviewKeyEvent = onPreviewKeyEvent,
-            onKeyEvent = onKeyEvent,
-            alwaysOnBottom = alwaysOnBottom,
-            content = content,
-        )
-    }
 
     /**
      * Opens a window driven by the AWT-free window API v2 clone
@@ -229,49 +165,6 @@ public fun interface NucleusDialogHost {
         onKeyEvent: (KeyEvent) -> Boolean,
         content: @Composable NucleusDecoratedDialogScope.() -> Unit,
     )
-
-    /**
-     * Opens a dialog driven by Compose Multiplatform 1.12's experimental
-     * dialog API v2.
-     *
-     * Default implementation converts [state] to v1 and calls [Dialog]
-     * so existing themed hosts keep their chrome. `minSize` / `maxSize`
-     * are v2-only and are dropped on that fallback. `requestScreen` /
-     * `screenId` are not applied on Tao (primary work area only).
-     */
-    @ExperimentalComposeUiApi
-    @Suppress("UnusedParameter")
-    @Composable
-    public fun Dialog(
-        onCloseRequest: () -> Unit,
-        state: DialogStateV2,
-        visible: Boolean,
-        title: String,
-        icon: Painter?,
-        resizable: Boolean,
-        enabled: Boolean,
-        focusable: Boolean,
-        minSize: DpSize,
-        maxSize: DpSize,
-        onPreviewKeyEvent: (KeyEvent) -> Boolean,
-        onKeyEvent: (KeyEvent) -> Boolean,
-        content: @Composable NucleusDecoratedDialogScope.() -> Unit,
-    ) {
-        val v1 = rememberSyncedDialogState(state, visible)
-        Dialog(
-            onCloseRequest = onCloseRequest,
-            state = v1,
-            visible = visible,
-            title = title,
-            icon = icon,
-            resizable = resizable,
-            enabled = enabled,
-            focusable = focusable,
-            onPreviewKeyEvent = onPreviewKeyEvent,
-            onKeyEvent = onKeyEvent,
-            content = content,
-        )
-    }
 
     /**
      * Opens a dialog driven by the AWT-free dialog API v2 clone
@@ -395,54 +288,6 @@ public object DefaultNucleusWindowHost : NucleusWindowHost {
         )
     }
 
-    @ExperimentalComposeUiApi
-    @Composable
-    override fun Window(
-        onCloseRequest: () -> Unit,
-        state: WindowStateV2,
-        visible: Boolean,
-        title: String,
-        icon: Painter?,
-        resizable: Boolean,
-        enabled: Boolean,
-        focusable: Boolean,
-        alwaysOnTop: Boolean,
-        undecorated: Boolean,
-        popupFor: NucleusWindow?,
-        nativePopupLayers: Boolean,
-        nativeContextMenu: Boolean,
-        hiddenFromDock: Boolean,
-        minSize: DpSize,
-        maxSize: DpSize,
-        onPreviewKeyEvent: (KeyEvent) -> Boolean,
-        onKeyEvent: (KeyEvent) -> Boolean,
-        alwaysOnBottom: Boolean,
-        content: @Composable NucleusDecoratedWindowScope.() -> Unit,
-    ) {
-        DecoratedWindow(
-            onCloseRequest = onCloseRequest,
-            state = state,
-            visible = visible,
-            title = title,
-            icon = icon,
-            resizable = resizable,
-            enabled = enabled,
-            focusable = focusable,
-            alwaysOnTop = alwaysOnTop,
-            undecorated = undecorated,
-            popupFor = popupFor,
-            nativePopupLayers = nativePopupLayers,
-            nativeContextMenu = nativeContextMenu,
-            hiddenFromDock = hiddenFromDock,
-            minSize = minSize,
-            maxSize = maxSize,
-            onPreviewKeyEvent = onPreviewKeyEvent,
-            onKeyEvent = onKeyEvent,
-            alwaysOnBottom = alwaysOnBottom,
-            content = content,
-        )
-    }
-
     /**
      * Full v2 path for the AWT-free clone: `DecoratedWindow` keeps `maxSize`
      * and hands the bridge the native window, so `requestScreen` and every
@@ -524,40 +369,6 @@ public object DefaultNucleusDialogHost : NucleusDialogHost {
             resizable = resizable,
             enabled = enabled,
             focusable = focusable,
-            onPreviewKeyEvent = onPreviewKeyEvent,
-            onKeyEvent = onKeyEvent,
-            content = content,
-        )
-    }
-
-    @ExperimentalComposeUiApi
-    @Composable
-    override fun Dialog(
-        onCloseRequest: () -> Unit,
-        state: DialogStateV2,
-        visible: Boolean,
-        title: String,
-        icon: Painter?,
-        resizable: Boolean,
-        enabled: Boolean,
-        focusable: Boolean,
-        minSize: DpSize,
-        maxSize: DpSize,
-        onPreviewKeyEvent: (KeyEvent) -> Boolean,
-        onKeyEvent: (KeyEvent) -> Boolean,
-        content: @Composable NucleusDecoratedDialogScope.() -> Unit,
-    ) {
-        DecoratedDialog(
-            onCloseRequest = onCloseRequest,
-            state = state,
-            visible = visible,
-            title = title,
-            icon = icon,
-            resizable = resizable,
-            enabled = enabled,
-            focusable = focusable,
-            minSize = minSize,
-            maxSize = maxSize,
             onPreviewKeyEvent = onPreviewKeyEvent,
             onKeyEvent = onKeyEvent,
             content = content,
@@ -686,104 +497,6 @@ public fun HostedDialog(
         resizable = resizable,
         enabled = enabled,
         focusable = focusable,
-        onPreviewKeyEvent = onPreviewKeyEvent,
-        onKeyEvent = onKeyEvent,
-        content = content,
-    )
-}
-
-/**
- * Opens a secondary window via [LocalNucleusWindowHost] using Compose
- * Multiplatform 1.12's experimental window API v2.
- *
- * `requestScreen` / `screenId` are not applied on Tao (primary work area
- * only).
- */
-@ExperimentalComposeUiApi
-@Suppress("FunctionNaming", "LongParameterList")
-@Composable
-public fun HostedWindow(
-    onCloseRequest: () -> Unit,
-    state: WindowStateV2,
-    visible: Boolean = true,
-    title: String = "",
-    icon: Painter? = null,
-    resizable: Boolean = true,
-    enabled: Boolean = true,
-    focusable: Boolean = true,
-    alwaysOnTop: Boolean = false,
-    undecorated: Boolean = false,
-    popupFor: NucleusWindow? = null,
-    nativePopupLayers: Boolean = false,
-    nativeContextMenu: Boolean = false,
-    hiddenFromDock: Boolean = false,
-    minSize: DpSize = DpSize.Unspecified,
-    maxSize: DpSize = DpSize.Unspecified,
-    onPreviewKeyEvent: (KeyEvent) -> Boolean = { false },
-    onKeyEvent: (KeyEvent) -> Boolean = { false },
-    alwaysOnBottom: Boolean = false,
-    content: @Composable NucleusDecoratedWindowScope.() -> Unit,
-) {
-    LocalNucleusWindowHost.current.Window(
-        onCloseRequest = onCloseRequest,
-        state = state,
-        visible = visible,
-        title = title,
-        icon = icon,
-        resizable = resizable,
-        enabled = enabled,
-        focusable = focusable,
-        alwaysOnTop = alwaysOnTop,
-        undecorated = undecorated,
-        popupFor = popupFor,
-        nativePopupLayers = nativePopupLayers,
-        nativeContextMenu = nativeContextMenu,
-        hiddenFromDock = hiddenFromDock,
-        minSize = minSize,
-        maxSize = maxSize,
-        onPreviewKeyEvent = onPreviewKeyEvent,
-        onKeyEvent = onKeyEvent,
-        alwaysOnBottom = alwaysOnBottom,
-        content = content,
-    )
-}
-
-/**
- * Opens a secondary dialog via [LocalNucleusDialogHost] using Compose
- * Multiplatform 1.12's experimental dialog API v2.
- *
- * `requestScreen` / `screenId` are not applied on Tao (primary work area
- * only).
- */
-@ExperimentalComposeUiApi
-@Suppress("FunctionNaming", "LongParameterList")
-@Composable
-public fun HostedDialog(
-    onCloseRequest: () -> Unit,
-    state: DialogStateV2,
-    visible: Boolean = true,
-    title: String = "",
-    icon: Painter? = null,
-    resizable: Boolean = false,
-    enabled: Boolean = true,
-    focusable: Boolean = true,
-    minSize: DpSize = DpSize.Unspecified,
-    maxSize: DpSize = DpSize.Unspecified,
-    onPreviewKeyEvent: (KeyEvent) -> Boolean = { false },
-    onKeyEvent: (KeyEvent) -> Boolean = { false },
-    content: @Composable NucleusDecoratedDialogScope.() -> Unit,
-) {
-    LocalNucleusDialogHost.current.Dialog(
-        onCloseRequest = onCloseRequest,
-        state = state,
-        visible = visible,
-        title = title,
-        icon = icon,
-        resizable = resizable,
-        enabled = enabled,
-        focusable = focusable,
-        minSize = minSize,
-        maxSize = maxSize,
         onPreviewKeyEvent = onPreviewKeyEvent,
         onKeyEvent = onKeyEvent,
         content = content,
